Refactor/reorganize native code
