+++ libs/native/libraries/build.gradle
@@ -19,7 +19,7 @@ configurations {
 }
 
 var zstdVersion = "1.5.7"
-var vecVersion = "1.0.28"
+var vecVersion = "1.0.29"
 
 repositories {
   exclusiveContent {
