+++ libs/simdvec/native/src/vec/headers/vec.h
@@ -35,56 +35,3 @@
     #error "This library is meant to be compiled with a C++ compiler"
 #endif
 
-EXPORT int vec_caps();
-
-EXPORT int32_t vec_dot7u(const int8_t* a, const int8_t* b, const int32_t dims);
-
-EXPORT void vec_dot7u_bulk(const int8_t* a, const int8_t* b, const int32_t dims, const int32_t count, f32_t* results);
-
-EXPORT void vec_dot7u_bulk_offsets(
-    const int8_t* a,
-    const int8_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results);
-
-EXPORT int32_t vec_sqr7u(const int8_t* a, const int8_t* b, const int32_t length);
-
-EXPORT void vec_sqr7u_bulk(const int8_t* a, const int8_t* b, const int32_t dims, const int32_t count, f32_t* results);
-
-EXPORT void vec_sqr7u_bulk_offsets(
-    const int8_t* a,
-    const int8_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results);
-
-EXPORT f32_t vec_dotf32(const f32_t* a, const f32_t* b, const int32_t elementCount);
-
-EXPORT void vec_dotf32_bulk(const f32_t* a, const f32_t* b, const int32_t dims, const int32_t count, f32_t* results);
-
-EXPORT void vec_dotf32_bulk_offsets(
-    const f32_t* a,
-    const f32_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results);
-
-EXPORT f32_t vec_sqrf32(const f32_t* a, const f32_t* b, const int32_t elementCount);
-
-EXPORT void vec_sqrf32_bulk(const f32_t* a, const f32_t* b, const int32_t dims, const int32_t count, f32_t* results);
-
-EXPORT void vec_sqrf32_bulk_offsets(
-    const f32_t* a,
-    const f32_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results);
